+++ src/roi/roiSignal.m
@@ -28,9 +28,12 @@ function [signal, xpts] = roiSignal(imStack, roiMask, sampleRate, bkgdWindow, me
     % History:
     %   22Aug2020 - SSP
     %   02Dec2020 - SSP - Added bkgd estimation options
+    %   19Dec2020 - SSP - Removed first frame from analysis
     % --------------------------------------------------------------------
     
-    xpts = 1/sampleRate : 1/sampleRate : size(imStack, 3)/sampleRate;
+    % Get xpts while accounting for throwing out first blank frame
+    xpts = 1/sampleRate : 1/sampleRate : (size(imStack, 3)+1)/sampleRate;
+    xpts(1) = [];
 
     [a, b] = find(roiMask == 1);
     signal = zeros(numel(a), size(imStack, 3));