+++ util/rmse.m
@@ -1,0 +1,3 @@
+function y = rmse(a, b)
+
+    y = sqrt(mean((a(:) - b(:))) .^ 2);